Cleaned up variable type compile errors

+++ include/pdcomserv/pdcomserv.h
@@ -68,7 +68,7 @@ struct variable* pdcomserv_signal(
         size_t n,                 /**< Element count.
                                    * If @dim != NULL, this is the number
                                    * elements in @dim */
-        const size_t dim[]        /**< Dimensions. If NULL, consider the
+        const unsigned int dim[] /**< Dimensions. If NULL, consider the
                                    * parameter to be a vector of length @n */
         );
 
@@ -123,7 +123,7 @@ struct variable *pdcomserv_parameter(
         void *addr,               /**< Parameter address */
         size_t n,                 /**< Element count.  If @dim != NULL, this
                                    * is the number elements in * @dim */
-        const size_t dim[],       /**< Dimensions. If NULL, consider the
+        const unsigned int dim[], /**< Dimensions. If NULL, consider the
                                    * parameter to be a vector of length @n */
         paramtrigger_t paramopy,  /**< Callback for updating the parameter
                                    * inside real time context */
