+++ src/Signal.cpp
@@ -43,7 +43,7 @@ Signal::Signal( Task *task,
         const char *path,
         enum si_datatype_t dtype,
         unsigned int ndims,
-        const size_t *dim):
+        const unsigned int *dim):
     Variable(path, dtype, ndims, dim),
     decimation(decimation), task(task) //,  main(main)
 {
