+++ src/Variable.h
@@ -39,7 +39,7 @@ class Variable {
         Variable( const char *path,
                 enum si_datatype_t dtype,
                 unsigned int ndims = 1,
-                const size_t dim[] = 0);
+                const unsigned int dim[] = 0);
 
         virtual ~Variable();
 
@@ -65,7 +65,7 @@ class Variable {
         size_t * const dim;
 
         static size_t getDTypeSize(enum si_datatype_t dtype);
-        static size_t getNElem( unsigned int dims, const size_t dim[]);
+        static size_t getNElem( unsigned int dims, const unsigned int dim[]);
 };
 
 }
